# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "English Upoe API",
    "version": "1.0.0",
    "description": "Dokumentasi API untuk English Upoe dengan session-based authentication"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/auth/status": {
      "get": {
        "summary": "Cek Status Login",
        "description": "Cek apakah user sudah login menggunakan Google OAuth.",
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "User sudah login",
            "content": {
              "application/json": {
                "example": {
                  "id": "123456",
                  "firstName": "John",
                  "lastName": "Doe",
                  "email": "johndoe@gmail.com"
                }
              }
            }
          },
          "401": {
            "description": "User belum login"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Logout Google OAuth",
        "description": "Menghapus sesi login Google",
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout berhasil"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Gunakan cookie session untuk autentikasi"
      }
    }
  }
}